backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
pydantic[email]
sqlalchemy>=2.0
aiomysql
pymysql
cryptography
python-jose[cryptography]
passlib[bcrypt]
reportlab